rename adding organizer form component
