+++ apps/site/src/app/admin/directors/Directors.tsx
@@ -8,7 +8,7 @@ import Box from "@cloudscape-design/components/box";
 import Cards from "@cloudscape-design/components/cards";
 import Header from "@cloudscape-design/components/header";
 
-import OrganizerInput from "./OrganizerInput";
+import OrganizerInput from "./AddOrganizer";
 import useOrganizers, { Organizer } from "@/lib/admin/useOrganizers";
 import UserContext from "@/lib/admin/UserContext";
 import { isDirector } from "@/lib/admin/authorization";
